requirements.txt
fastapi
pydantic
uvicorn
streamlit
streamlit-authenticator
pyyaml